requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
langchain-anthropic
langchain-deepseek
python-dotenv
pydantic
duckduckgo-search